pyproject.toml
[project]
name = "today-in-tma"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-frontmatter>=1.1.0",
]
